Added "MappedIntegerBuffer" Class and added dirty flag support to "IntegerBuffer"
I've added an extension to IntegerBuffer called "MappedIntegerBuffer"
which encapsulates scaling of the value to a given range, and only sets
the dirty flag is this scaled value has changed.  e.g. Mapping compass
heading (0-65535) onto an LCD readout (1-360).

This required an accessor change to IntegerBuffer and I also modified
onDcsBiosWrite to set the dirty flag appropriately.