Int16Buffer: fix overzealous spurious update suppression logic
The Int16Buffer class will only mark the data as dirty
(i.e. have hasUpdatedData() return true) when the received data value
is actually different from the last value that was seen.

The previous logic failed to take into account the uninitialized state,
where the (zero) value stored in the data member variable was not the
result of a previous update from DCS-BIOS, but the default after
program startup.